Add MenageIntegration tab
