+++ resources/translations/nl.json
@@ -133,7 +133,6 @@
   },
   "details": {
     "save": "Wijzigingen opslaan",
-    "delete": "Integratie verwijderen",
     "basic_info": {
       "title": "Basisgegevens",
       "name": "Integratienaam",
@@ -184,6 +183,15 @@
       "environment": "Omgeving",
       "logout": "Logout URL",
       "dialog": "Wil je deze URL echt verwijderen?"
+    },
+    "menage_account": {
+      "title": "Integratiebeheer",
+      "delete": {
+        "title": "Integratie verwijderen",
+        "description": "Ben je zeker dat je je integratie wilt verwijderen? Nadat je hebt gekozen om je integratie te verwijderen, zal deze niet langer toegankelijk zijn en kan de verwijdering niet ongedaan gemaakt worden."
+      },
+
+      "action": "Verwijderen"
     }
   },
   "support": {
